+++ test/mocks_test.go
@@ -5,6 +5,7 @@ import (
 	"github.com/raxigan/pcfy-my-mac/install"
 	"os"
 	"strings"
+	"time"
 )
 
 type MockCommander struct {
@@ -33,4 +34,12 @@ func (c *MockCommander) Run(command string) {
 
 func (c *MockCommander) Exists(command string) bool {
 	return true
+}
+
+type FakeTimeProvider struct {
+}
+
+func (tp FakeTimeProvider) Now() time.Time {
+	parse, _ := time.Parse("2006-01-02 15:04:05", "2023-09-27 12:30:00")
+	return parse
 }
